# Generic/LowerCaseConstant: improve performance