--- a/src/Standards/Generic/Sniffs/PHP/LowerCaseConstantSniff.php
+++ b/src/Standards/Generic/Sniffs/PHP/LowerCaseConstantSniff.php
@@ -37,6 +37,29 @@ class LowerCaseConstantSniff implements Sniff
         T_NULL  => T_NULL,
     ];
 
+    /**
+     * Token types which can be encountered in a property type declaration.
+     *
+     * @var array<int|string, int|string>
+     */
+    private $propertyTypeTokens = [
+        T_CALLABLE             => T_CALLABLE,
+        T_SELF                 => T_SELF,
+        T_PARENT               => T_PARENT,
+        T_FALSE                => T_FALSE,
+        T_TRUE                 => T_TRUE,
+        T_NULL                 => T_NULL,
+        T_STRING               => T_STRING,
+        T_NAME_QUALIFIED       => T_NAME_QUALIFIED,
+        T_NAME_FULLY_QUALIFIED => T_NAME_FULLY_QUALIFIED,
+        T_NAME_RELATIVE        => T_NAME_RELATIVE,
+        T_NS_SEPARATOR         => T_NS_SEPARATOR,
+        T_NAMESPACE            => T_NAMESPACE,
+        T_TYPE_UNION           => T_TYPE_UNION,
+        T_TYPE_INTERSECTION    => T_TYPE_INTERSECTION,
+        T_NULLABLE             => T_NULLABLE,
+    ];
+
 
     /**
      * Returns an array of tokens this test wants to listen for.
@@ -47,7 +70,13 @@ class LowerCaseConstantSniff implements Sniff
     {
         $targets = $this->targets;
 
-        // Register function keywords to filter out type declarations.
+        // Register scope modifiers to filter out property type declarations.
+        $targets  += Tokens::$scopeModifiers;
+        $targets[] = T_VAR;
+        $targets[] = T_STATIC;
+        $targets[] = T_READONLY;
+
+        // Register function keywords to filter out param/return type declarations.
         $targets[] = T_FUNCTION;
         $targets[] = T_CLOSURE;
         $targets[] = T_FN;
@@ -64,11 +93,42 @@ class LowerCaseConstantSniff implements Sniff
      * @param int                         $stackPtr  The position of the current token in the
      *                                               stack passed in $tokens.
      *
-     * @return void|int
+     * @return void|int Optionally returns a stack pointer. The sniff will not be
+     *                  called again on the current file until the returned stack
+     *                  pointer is reached.
      */
     public function process(File $phpcsFile, $stackPtr)
     {
         $tokens = $phpcsFile->getTokens();
+
+        /*
+         * Skip over type declarations for properties.
+         *
+         * Note: for other uses of the visibility modifiers (functions, constants, trait use),
+         * nothing relevant will be skipped as the next non-empty token will be an "non-skippable"
+         * one.
+         * Functions are handled separately below (and then skip to their scope opener), so
+         * this should also not cause any confusion for constructor property promotion.
+         *
+         * For other uses of the "static" keyword, it also shouldn't be problematic as the only
+         * time the next non-empty token will be a "skippable" token will be in return type
+         * declarations, in which case, it is correct to skip over them.
+         */
+
+        if (isset(Tokens::$scopeModifiers[$tokens[$stackPtr]['code']]) === true
+            || $tokens[$stackPtr]['code'] === T_VAR
+            || $tokens[$stackPtr]['code'] === T_STATIC
+            || $tokens[$stackPtr]['code'] === T_READONLY
+        ) {
+            $skipOver = (Tokens::$emptyTokens + $this->propertyTypeTokens);
+            $skipTo   = $phpcsFile->findNext($skipOver, ($stackPtr + 1), null, true);
+            if ($skipTo !== false) {
+                return $skipTo;
+            }
+
+            // If we're at the end of the file, just return.
+            return;
+        }
 
         // Handle function declarations separately as they may contain the keywords in type declarations.
         if ($tokens[$stackPtr]['code'] === T_FUNCTION
@@ -79,9 +139,15 @@ class LowerCaseConstantSniff implements Sniff
                 return;
             }
 
+            // Make sure to skip over return type declarations.
             $end = $tokens[$stackPtr]['parenthesis_closer'];
             if (isset($tokens[$stackPtr]['scope_opener']) === true) {
                 $end = $tokens[$stackPtr]['scope_opener'];
+            } else {
+                $skipTo = $phpcsFile->findNext([T_SEMICOLON, T_OPEN_CURLY_BRACKET], ($end + 1), null, false, null, true);
+                if ($skipTo !== false) {
+                    $end = $skipTo;
+                }
             }
 
             // Do a quick check if any of the targets exist in the declaration.
@@ -113,21 +179,6 @@ class LowerCaseConstantSniff implements Sniff
             // Skip over return type declarations.
             return $end;
         }//end if
-
-        // Handle property declarations separately as they may contain the keywords in type declarations.
-        if (isset($tokens[$stackPtr]['conditions']) === true) {
-            $conditions    = $tokens[$stackPtr]['conditions'];
-            $lastCondition = end($conditions);
-            if (isset(Tokens::$ooScopeTokens[$lastCondition]) === true) {
-                // This can only be an OO constant or property declaration as methods are handled above.
-                $equals = $phpcsFile->findPrevious(T_EQUAL, ($stackPtr - 1), null, false, null, true);
-                if ($equals !== false) {
-                    $this->processConstant($phpcsFile, $stackPtr);
-                }
-
-                return;
-            }
-        }
 
         // Handle everything else.
         $this->processConstant($phpcsFile, $stackPtr);