--- a/src/Standards/Generic/Tests/PHP/LowerCaseConstantUnitTest.php
+++ b/src/Standards/Generic/Tests/PHP/LowerCaseConstantUnitTest.php
@@ -53,6 +53,12 @@ class LowerCaseConstantUnitTest extends AbstractSniffUnitTest
                 100 => 2,
                 104 => 1,
                 108 => 1,
+                118 => 1,
+                119 => 1,
+                120 => 1,
+                121 => 1,
+                125 => 1,
+                129 => 1,
             ];
 
         case 'LowerCaseConstantUnitTest.js':